# Backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Unora mobile application",
        "title": "Unora Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Unora-app/Backend",
            "email": "support@unora.app"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/Unora-app/Backend"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/hobbies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all hobbies for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hobbies"
                ],
                "summary": "Get user hobbies",
                "responses": {
                    "200": {
                        "description": "User hobbies",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.HobbyResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a hobby to user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hobbies"
                ],
                "summary": "Add hobby",
                "parameters": [
                    {
                        "description": "Hobby data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.AddHobbyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Hobby added",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.HobbyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or hobby already added",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/hobbies/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a hobby from user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hobbies"
                ],
                "summary": "Delete hobby",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hobby ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hobby deleted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/hobby-options": {
            "get": {
                "description": "Get all available hobby options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hobbies"
                ],
                "summary": "Get hobby options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hobby options",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.HobbyOptionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/photos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all photos for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Get user photos",
                "responses": {
                    "200": {
                        "description": "User photos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.PhotoResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a photo after successful upload to storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Register uploaded photo",
                "parameters": [
                    {
                        "description": "Photo registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.RegisterPhotoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Photo registered",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.PhotoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or max photos reached",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/photos/upload-url": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a presigned URL for uploading a photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Get photo upload URL",
                "parameters": [
                    {
                        "description": "Upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.UploadPhotoURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload URL",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.UploadPhotoURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/photos/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a photo from user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Delete photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo deleted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Photo not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/profile/onboarding-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the onboarding completion status for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get onboarding status",
                "responses": {
                    "200": {
                        "description": "Onboarding status",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.OnboardingStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_UnoraApp_be_internal_profile_dto.HobbyResponse": {
            "description": "User's selected hobby",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "displayOrder": {
                    "type": "integer",
                    "example": 1
                },
                "hobbyOption": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_internal_profile_dto.HobbyOptionResponse"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "microDescription": {
                    "type": "string",
                    "example": "I love landscape photography"
                }
            }
        },
...
        "github_com_UnoraApp_be_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}